requirements.txt
Flask
gunicorn
pillow
requests
flask-cors
